extracts common descriptions
